\name{hse-internal}
\alias{grad}
\alias{hess}
\title{Internal hse functions.}
\description{
  Internal hse functions.
}
\usage{
grad(x, alpha, beta, ntop, zeta)
hess(alpha, beta, ntop, zeta, ndata)
}
\details{
  These functions are auxiliary and are not intended to be called
  by the user.
}
\keyword{internal}
